+++ History.md
@@ -3,6 +3,10 @@
 
 This incorporates all changes after 4.17.1 up to 4.17.2.
 
+  * deps: body-parser@2.0.0-beta.1
+    - `req.body` is no longer always initialized to `{}`
+    - `urlencoded` parser now defaults `extended` to `false`
+    - Use `on-finished` to determine when body read
   * deps: router@2.0.0-beta.1
     - Add new `?`, `*`, and `+` parameter modifiers
     - Internalize private `router.process_params` method